Storable: add Unsafe modules
